+++ src/sll/include/sll/io.h
@@ -26,6 +26,19 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_load_assembly(sll_file_t* rf,sl
 
 /**
  * \flags check_output func
+ * \name sll_load_bundle
+ * \group io
+ * \desc Docs!
+ * \arg sll_file_t* rf
+ * \arg sll_bundle_t* b
+ * \ret sll_bool_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_load_bundle(sll_file_t* rf,sll_bundle_t* b);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_load_compiled_node
  * \group io
  * \desc Docs!
@@ -43,9 +56,21 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_load_compiled_node(sll_file_t* 
  * \group io
  * \desc Docs!
  * \arg sll_file_t* wf
- * \arg const sll_assembly_data_t* o
+ * \arg const sll_assembly_data_t* a_dt
  */
-__SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t* o);
+__SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t* a_dt);
+
+
+
+/**
+ * \flags func
+ * \name sll_write_bundle
+ * \group io
+ * \desc Docs!
+ * \arg sll_file_t* wf
+ * \arg const sll_bundle_t* b
+ */
+__SLL_EXTERNAL void sll_write_bundle(sll_file_t* wf,const sll_bundle_t* b);